requirements.txt
setuptools>=69.0.3
wheel>=0.42.0
Django==5.0.10
django-admin-interface==0.28.0
django-bootstrap-form==3.4
django-bootstrap-forms==0.1
django-colorfield==0.10.0
django-crispy-forms==2.3
django-filter==24.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-widget-tweaks==1.5.0
django-allauth==0.61.1
django-phonenumber-field[phonenumbers]==7.3.0
Pillow==11.2.1
djangorestframework==3.15.1
requests==2.32.2
crispy-bootstrap5==2025.4 
